refactor(backend): modularize prompt cache conversations
