+++ docs/specs/README.md
@@ -166,3 +166,4 @@
 | ts6qp | account-pool 前端类型债清偿 | 已完成 | `ts6qp-account-pool-typing-debt/SPEC.md` | 2026-04-12 | normal-flow / production `@ts-nocheck` removed from account-pool pages / `UpstreamAccounts` helper extraction + Create strict typing landed / local lint+test+build+storybook green / visual evidence archived |
 | 9vau7 | 后端结构债双 PR 快车道 | 已完成 | `9vau7-backend-structure-dual-pr-followup/SPEC.md` | 2026-04-12 | fast-track / PR1=`proxy+upstream_accounts/routing` 已合并 / PR2=`api+router builders+api-read smoke` 本地 cargo fmt+check+test 与 shared-testbox api-read+proxy-parallel+raw smoke 通过 / merge+cleanup |
 | g7n33 | 后端 archive 结构收敛 | 已实现，待 PR / CI / review-proof 收敛 | `g7n33-backend-archive-structure-convergence/SPEC.md` | 2026-04-12 | fast-track / archive-only follow-up / split `src/maintenance/archive.rs` into real submodules / local cargo fmt+check+test + shared-testbox raw/api-read smoke green / merge+cleanup |
+| n78zb | 后端 prompt-cache conversations 结构收敛 | 已实现，待 PR / CI / review-proof 收敛 | `n78zb-backend-prompt-cache-conversations-structure/SPEC.md` | 2026-04-12 | fast-track / prompt-cache-only follow-up / split `src/api/slices/prompt_cache_and_timeseries/prompt_cache_conversations.rs` into real submodules / local cargo fmt+check+test green / shared-testbox api-read smoke green / merge+cleanup |
