+++ src/api/slices/prompt_cache_and_timeseries/mod.rs
@@ -10,9 +10,8 @@ pub(crate) use prompt_cache_and_timeseries_summary_queries::*;
 #[path = "forward_proxy_stats.rs"]
 mod prompt_cache_and_timeseries_forward_proxy_stats;
 pub(crate) use prompt_cache_and_timeseries_forward_proxy_stats::*;
-#[path = "prompt_cache_conversations.rs"]
-mod prompt_cache_and_timeseries_prompt_cache_conversations;
-pub(crate) use prompt_cache_and_timeseries_prompt_cache_conversations::*;
+mod prompt_cache_conversations;
+pub(crate) use prompt_cache_conversations::*;
 #[path = "timeseries.rs"]
 mod prompt_cache_and_timeseries_timeseries;
 pub(crate) use prompt_cache_and_timeseries_timeseries::*;
